Move etcd snapshot code into separate file
